+++ dev4press/core/mailer/Detection.php
@@ -374,7 +374,7 @@ class Detection {
 				'source' => 'coreSecurity',
 				'label'  => _x( 'Instant Notification', 'Email Detection Type', 'd4plib' ),
 			),
-			'coresecurity-user-notification'                      => array(
+			'coresecurity-user-notification'                         => array(
 				'filter' => 'coresecurity-user-notification-email-subject',
 				'source' => 'coreSecurity',
 				'label'  => _x( 'User Notification', 'Email Detection Type', 'd4plib' ),