+++ changelog.md
@@ -2,10 +2,12 @@
 
 ## Changelog
 
-### Version 5.4 - 2025.04.05
+### Version 5.4 - 2025.05.24
 
 * New: additional free flag in the License
 * New: show admin notices when license is not activated
+* New: updated license validation on save
+* New: improved weekly license validation
 
 ### Version 5.3 - 2025.01.26